moved kill_process and dir_manager to helpers
